Refactor package structure (#30)
Driver interface has been replaced with hypp/js and hypp/jsd package. The hypp/js package is a replacement for the syscall/js package. It can be used without directly depending on syscall/js. The hypp/jsd package uses syscall/js to implement the hypp/js driver and registers it in the hypp/js package on init. This is similar to how drivers are registered in database/sql.

* Refactor events

* Add function to remove child and child's events

* Update examples

* Update package dependency graph

* Add package name to hypp/window error messages

* Remove Render type

* Remove Map type

* Export ValidateHProps functions

* Rename hPropsKeys function to mergeAndSortHPropsKeys

* Refactor html Render and Parse functions

* Test validate function doesn't panic for short keys

* Add documentation

* Do not access character of key if out of bounds

* Make sure 'events' is never undefined

* Validate HProps returns error

* Remove duplicate validation

* Upgrade dependencies